requirements/base.txt
Django==2.0.5
psycopg2==2.7.4
pytz==2018.4
Unipath==1.1
South==1.0.2
social-auth-app-django==2.1.0
djangorestframework==3.8.2
ply==3.11